Made the project code more modular
